Fix
The Emails from when a campaign is activated and when notify if there is capacity for the event, now send a url that let the person who receive the mail can click there and redirect them into the record.

I think we need Team 1 set Validation Rules into some fields of the objects for a much better control.