feat: event-registration-flows
* feat: First event-registration-flows

I have implemented three flows to streamline event registration. The first flow sends email updates when a campaign’s status changes to 'In Progress.' The second ensures that registrations are accepted only if the event has capacity, creating a confirmation task for approved registrants. The third manages the waitlist, notifying participants when spots become available.
I have configured email locally, but they are not sending even if they grab data correctly. I am still looking for the issue; however, I tested the flows in debug mode, and they work as expected.

* Fix names

There was a problem at the moment to create the names of the registration form an confirmation task. Now on the first one show the event and the name of the student and in the last one just the name of the student

* Fix

The Emails from when a campaign is activated and when notify if there is capacity for the event, now send a url that let the person who receive the mail can click there and redirect them into the record.

I think we need Team 1 set Validation Rules into some fields of the objects for a much better control.

* Delete force-app/main/default/flows/Flow_send_Email_when_Status_is_Accepted.flow-meta.xml

Uploaded by mistake

* Delete force-app/main/default/flows/Flow_Waitlist.flow-meta.xml

I uploaded a new variation