feat: pre event reminder
* Feature pre event reminder and fixes

I revamp the PreEvent_Reminder and make fixes on the flow of waitlist management where the person can become a campaign member putting the flow to wait. I change the validate to capacity to the case when a record is created or updated, that allows to accept Registration form made on the web page that Jonatan did.

I made an new custom field for campaigns, the reason was that I need to confirm that the event is happening the day the flow is running. I didn´t load the permissions for the new profiles, but Team 1 can do it :)

For some flows I might need configure Org Wide Email Address and bring a new email into the flow given that some flows are scheduled and cannot grab email from the user who activate the flow.

* chore: manifiest update

---------

Co-authored-by: Jonatan <jsantanamay@gmail.com>