build: the firts releases
* feat: First custom and standard objects model created. Standard, Custom and corresponding Validation Rules are included.

* feat-1

* feat-1-1

* doc: package.xml added

* feat: Added the wailist field for the Registration Form Object and removed Case non-existent fields references into profiles.

* feat-2

* feat-3

* feat: tabs added

* feat: event-registration-flows

* feat: First event-registration-flows

I have implemented three flows to streamline event registration. The first flow sends email updates when a campaign’s status changes to 'In Progress.' The second ensures that registrations are accepted only if the event has capacity, creating a confirmation task for approved registrants. The third manages the waitlist, notifying participants when spots become available.
I have configured email locally, but they are not sending even if they grab data correctly. I am still looking for the issue; however, I tested the flows in debug mode, and they work as expected.

* Fix names

There was a problem at the moment to create the names of the registration form an confirmation task. Now on the first one show the event and the name of the student and in the last one just the name of the student

* Fix

The Emails from when a campaign is activated and when notify if there is capacity for the event, now send a url that let the person who receive the mail can click there and redirect them into the record.

I think we need Team 1 set Validation Rules into some fields of the objects for a much better control.

* Delete force-app/main/default/flows/Flow_send_Email_when_Status_is_Accepted.flow-meta.xml

Uploaded by mistake

* Delete force-app/main/default/flows/Flow_Waitlist.flow-meta.xml

I uploaded a new variation

* feat: Adding details to layouts Institutions or Organizations AO and EC

* Added layouts for institutions or organizations (AO and EC)

* Adding details to layouts Institutions or Organizations AO and EC

---------

Co-authored-by: IndraAst <blueasterope@gmail.com>

* feat: Post-event attendance update flow 

* Feat Flow PostEvent AttendanceUpdate

* Remove Extra file

* Bugs fixed

* fix: Send individual emails to all campaign members

* feat: digital experience metadata upload

* feat: lwc to creat a RegisterForm added

* feat: digital experience metadata upload

* feat: test controllers added

* feat: test lwc added

* feat: metadata newtwork

* feat: Added layouts "Registration Form" and "Confirmation task"

* feat: :lipstick: Added layouts "Registration Form" and "Confirmation Task"

Registration Form (Custom Object) - Added layout with all required fields. - Added “Submit for Approval” button Confirmation Task (Custom Object) - Added two sections: General and Registration Form - Added all fields to the layout. - Added “Submit Survey” button.

* feat: :lipstick: Modify layouts Confirmation Task and Registration Form

Added 2 layouts for each profile (Event Coordinator) and (Admission Officer) for Customs Objects.

* chore: layout manifiest update

* feat: pre event reminder

* Feature pre event reminder and fixes

I revamp the PreEvent_Reminder and make fixes on the flow of waitlist management where the person can become a campaign member putting the flow to wait. I change the validate to capacity to the case when a record is created or updated, that allows to accept Registration form made on the web page that Jonatan did.

I made an new custom field for campaigns, the reason was that I need to confirm that the event is happening the day the flow is running. I didn´t load the permissions for the new profiles, but Team 1 can do it :)

For some flows I might need configure Org Wide Email Address and bring a new email into the flow given that some flows are scheduled and cannot grab email from the user who activate the flow.

* chore: manifiest update

---------

Co-authored-by: Jonatan <jsantanamay@gmail.com>

* feat: Added profiles permissions and new student profile

* feat-3

* feat-4

* New custom object, reviewed permissions.

* Admin profile and tab added

* chore: manifiest updated

* feat: email, reports and dashboard templates ready 

* email, reports and dashboard templates ready

* PR fixed

* dashboard fix for pr

* docs: manifiest added

---------

Co-authored-by: Jonatan <jsantanamay@gmail.com>
Co-authored-by: JonatanSantana <72703077+JonatanJSM@users.noreply.github.com>

* feat: Update flowPostEvent attendance

* feat: account and contact layouts added

* Cambios Layout Perfiles 28Nov24

* cambios solicitados para fileld active

---------

Co-authored-by: Jonatan <jsantanamay@gmail.com>

---------

Co-authored-by: Jesús Méndez <78876044+Jsvela749@users.noreply.github.com>
Co-authored-by: Andrés Mena Salazar <56236562+sajeme@users.noreply.github.com>
Co-authored-by: Chelles <112115187+IndraAst@users.noreply.github.com>
Co-authored-by: IndraAst <blueasterope@gmail.com>
Co-authored-by: Omar Cauich <omar231120@hotmail.com>
Co-authored-by: Daniel García <63012886+Daniel-Garcia-C@users.noreply.github.com>
Co-authored-by: Juan Osvaldo <82409690+ThuPapuha@users.noreply.github.com>
Co-authored-by: ArmandoHdz02 <armahv@gmail.com>